BUGFIX: CC'ed on final game allowed to view different accounts
When people got CC'ed on the final email, they also got copies of everyones
email and the link to the game in that email. That way they could go to the
user page of a different user.

Removed CC option and will just send out two emails from now on. The first one to
all players (reply all possible) and the second one with the personal link.