BUGFIX: output call only after game started; move after game comments to right menu
the re/contra, etc. box showed up after a user finished the setup, changed it so that
it now only shows up after the game has started.

The input form for comments after a game has finished was still in the card area, moved
it to the right game menu.