BUGFIX: couldn't restart a finished game; force reload of CSS
problem was that everything on the page was in a form which is now
submitted to the game page, but for new games, we need to call the newgame.php
page. moved the button out of the side and into its own div structure now to fix it.

also renamed the css file to force a reload

Signed-off-by: Arun Persaud <arun@nubati.net>