BUGFIX: setting the gametype gave away information in case someone wanted to play solo.
setting the gametype to what was stored in the DB sorted the hands of everyone according to the gametype. This showed for example if someone wanted to play a solo, since then everybodys cards where ordered like that... changed it now, so that for the pre-game phase the cards are ordered for normal games.

Signed-off-by: Arun Persaud <arun@nubati.net>