BUGFIX: user menu didn't list all games where it is your turn
games where you need to decide if you want to play should be listed
in the menu on the left. The problem is that once somebody moved they
won't show up in the list anymore, since from that point on the server
keeps track of who needs to move next.

Signed-off-by: Arun Persaud <arun@nubati.net>