fixed showing re/contra call at beginning of game; allowed pre-game comments; DB upgrade
the problem with re/contra at the beginning of a game was that no playid was set
yet, so we couldn't record when the call was made. Added two special playid (-1 and -2),
for pre-game comments and re/contra calls before the first card is played.
For this we need to record the gameid in the comment table => DB update