fixed margin collapse in .main (gray bar below header)
the h4 element has a default top margin, but the .user and .main diff didn't have any.
Because of collapsing margins they were also assigned the same margin which then created the gray bar. This doesn't happen if the parent element has non-zero padding, so I added a 1px top and bottom padding to .main, which fixed the problem