BUGFIX: fixed selection of spade and heart solo
two typos messed up the solo selction, e.g. the gametype was not the same as the one given in the
enum in the database. Also added an additional check before we set them in the database, so that people
can mess things up when playing with the url directly.